+++ src/ProSidebar/MenuItem.tsx
@@ -5,10 +5,11 @@ export interface Props {
   className?: string;
   icon?: React.ReactNode;
   active?: boolean;
+  suffix?: React.ReactNode;
 }
 
 const MenuItem: React.ForwardRefRenderFunction<unknown, Props> = (
-  { children, className, icon, active, ...rest },
+  { children, className, icon, active, suffix, ...rest },
   ref,
 ) => {
   const menuItemRef: LegacyRef<HTMLLIElement> = (ref as any) || React.createRef<HTMLLIElement>();
@@ -23,6 +24,7 @@ const MenuItem: React.ForwardRefRenderFunction<unknown, Props> = (
         ) : null}
 
         <span className="pro-item-content">{children}</span>
+        {suffix ? <span className="suffix-wrapper">{suffix}</span> : null}
       </div>
     </li>
   );