+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add suffix prop to menuItem and submenu
+
 ## [0.2.0] - 2020-05-08
 
 ### Added